app/requirements.txt
torch==2.2.2
torchvision==0.17.2
pillow==10.4.0
fastapi==0.115.0
jinja2==3.1.4
numpy==1.26.4
pip==24.2
pydantic==2.9.2
uvicorn==0.31.0
wheel==0.41.2
annotated-types==0.7.0
starlette==0.38.6
anyio==4.6.0
click==8.1.7
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
pydantic-core==2.23.4
python-multipart==0.0.12
sniffio==1.3.1
typing-extensions==4.12.2
